Refactor: 카테고리 삭제 처리 수정

+++ src/main/java/techit/gongsimchae/domain/groupbuying/category/service/CategoryService.java
@@ -1,12 +1,16 @@
 package techit.gongsimchae.domain.groupbuying.category.service;
 
+import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
+
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import techit.gongsimchae.domain.groupbuying.category.dto.CategoryReqDtoWeb;
+import techit.gongsimchae.domain.groupbuying.category.dto.CategoryWithItemDto;
 import techit.gongsimchae.domain.groupbuying.category.entity.Category;
 import techit.gongsimchae.domain.groupbuying.category.repository.CategoryRepository;
 import techit.gongsimchae.domain.groupbuying.event.entity.Event;
@@ -14,6 +18,7 @@ import techit.gongsimchae.domain.groupbuying.event.repository.EventRepository;
 import techit.gongsimchae.domain.groupbuying.eventcategory.entity.EventCategory;
 import techit.gongsimchae.domain.groupbuying.eventcategory.repository.EventCategoryRepository;
 import techit.gongsimchae.domain.groupbuying.item.entity.Item;
+import techit.gongsimchae.domain.groupbuying.item.entity.ItemStatus;
 import techit.gongsimchae.domain.groupbuying.item.repository.ItemRepository;
 import techit.gongsimchae.global.exception.CustomWebException;
 import techit.gongsimchae.global.message.ErrorMessage;
@@ -68,7 +73,20 @@ public class CategoryService {
         Category category = categoryRepository.findById(categoryId).orElseThrow(
                 () -> new CustomWebException(ErrorMessage.CATEGORY_NOT_FOUND)
         );
+
+        if(itemRepository.existsByCategoryAndItemStatus(category, ItemStatus.공동구매_진행중)){
+            throw new CustomWebException(ErrorMessage.CANNOT_DELETE_CATEGORY_WITH_ACTIVE_ITEM);
+        }
+
+        List<Item> items = itemRepository.findAllByCategoryAndDeleteStatus(category, 0);
+        items.forEach(Item::markAsDeleted);
+        itemRepository.saveAll(items);
+
+
+
         category.delete();
+        categoryRepository.save(category); // 카테고리 저장
+
     }
 
     public void restoreCategory(CategoryReqDtoWeb categoryDtoWeb) {
